Bug #21761867: ASSERTION `TYPE_ARG == MAYBE_KEY ||                TYPE_ARG == IMPOSSIBLE' FAILED.
Issue:
-----
When two SEL_ARG predicates are "And-ed" together, i.e.
combined into a common AND expression, the case when either
predicate is of the type ALWAYS is not handled. ALWAYS
flag represents a predicate that is unable to reject any
rows. When key_and tries to clone a SEL_ARG object of the
type ALWAYS, it results in a failed assertion.

SOLUTION:
---------
If either SEL_ARG type is ALWAYS, return the other key.
Essentially, when one of the keys in an AND-condition is
always true, the result will depend on the rest of the
keys.