Description: 1. Hint commentaries are designed to behave like regular /* ... */ commentaries:   SELECT /*+ " */ 1; -- a valid SQL query However, the mysql client program waits for the closing doublequote character. 2. There is a similar issue with statement delimiters like ';', ''', '`'
Analysis:
This regression is introduced by the changes of wl#8016. When we are parsing the input string we need to consider '+' as hint character. Which is not handled.

Fix: Introduced a hint flag for ss_comment and while parsing if we come across "+" we are making ss_comment as hint(SSC_HINT) and reseting it when we came across '*/'.