Bug#21682231 HANDLE_FATAL_SIGNAL (SIG=11) IN FREE_ROOT FROM TEST_QUICK_SELECT
Limit the effective maximal size for the system variables

  range_alloc_block_size    and
  query_alloc_block_size

to 2^32-1, adjusted to 2^32-1024, or 4294966272 see below.

Without this fix, when passing in a larger than 32 bits value on a 64
bits non-Windows system[1] as a start-up parameter, there is no range
checking and an overflow ensues when the ulong quantity is passed as a
uint to init_sql_alloc, leading to a crash (zero is passed in).

[1] The system variable, THD::variables.range_alloc_block_size has
type ulong.

Btw, this can't happen on Windows since ulong is 32 bits on that
platform.

The patch also changes the signature of init_sql_alloc to use size_t
instead of uint, more correct since the underlying call to
init_alloc_root takes a size_t type as well (future-proofing).

The documentation should be updated; it states that the maximal value
for this variable on a 64 bits platform is 18446744073709547520.  It
should now be UINT_MAX32 for all platforms (rounded off to multiple of
block_size: 4294966272 (2^32-1024) for both range_alloc_block_size and
query_alloc_block_size).

Tests adjusted.

This patch also fixes 21745936, 21795158 and 21601361.