Bug#21613615 GCOLS: ASSERTION FAILED: !TABLE || (!TABLE->READ_SET || BITMAP_IS_SET
Taking this table:
CREATE TABLE v (
a INT,
c INT,
b CHAR(2) GENERATED ALWAYS AS (a IN (1)) VIRTUAL,
KEY(c,b(1)));
and query:
SELECT (SELECT MAX(c) FROM v);

opt_sum_query() optimizes the query, decides to read the MAX(c) using
the index over (c,b(1)). It first does this:
          // Set bits for user-defined parts of key
          table->mark_columns_used_by_index_no_reset(ref.key, table->read_set);
which puts "c" and "b" in read_set (it's actually tmp_set).
mark_columns_used_by_index_no_reset() adds "b" but doesn't add
dependencies of "b", which looks correct as otherwise it would force a
read of the data row and make any index on "b" be useless.
Then in the index read, update_generated_read_fields() is called,
which calls index_contains_this_virtual_gcol() which says that "b" is
*not* in the index; it's because the index is on a prefix of "b",
(key_part->field isn't equal to vfield, it's a forged field based on
"b" but with a field_length of 1). It is also correct, "b" isn't in
the index, only a prefix is.
So update_generated_read_fields() wants to calculate "b", thus needs
"a", which isn't in read_set, and it fails.

Fix: put in this temporary read_set only columns which are necessary
for this index read. They are the columns involved in
'WHERE col=const' and the aggregated one.