Bug#21796092 - ASSERT IN RECOVERY, COMPRESSION::DESERIALIZE, MEMCMP(SRC + 16 + 4, SRC + (HEADE
Problem:
--------
The assertion checks LSN from page header matches to the one stored
in page trailer.

When there is torn write, the LSN debug check will fail

Fix:
----
If the page is from double-write buffer, it could be a torn page and
the LSN check will fail. So we skip this debug check for pages from
double-write buffer.

Approved by Sunny over IM.