Bug#21828321: JSON FUNCS CALL DBUG_ABORT OR EXIT() ON WINDOWS!
If LEAST or GREATEST is called with only JSON arguments, and the
result of the call is passed as an argument to a JSON function, an
error is raised in release builds and an assert failure in debug
builds.

The LEAST and GREATEST functions don't have any special handling of
JSON values currently. JSON arguments are handled as if they were
strings, and the result is returned as a string. If all the arguments
have type JSON, however, the return type of the function will be set
to MYSQL_TYPE_JSON. If the result is passed as an argument to a JSON
function, the JSON function will attempt to call val_json() on the
result, since it appears to be a JSON value. Since the corresponding
Item subclass (Item_func_min_max) does not override Item::val_json(),
an error or assert failure is raised when val_json() is called.

The fix changes the return type of the function call from JSON to
string in the case where all arguments are JSON values. This way, the
receiving JSON function is not led to believe it's a JSON value and
does not call val_json() on it.

This patch also makes LEAST and GREATEST raise a "not supported yet"
warning if one of the arguments is JSON. This is meant as a
notification that these operators have not yet been updated to use the
JSON comparator, and may therefore give other results than expected.

The warning is raised by BETWEEN and IN as well, since those operators
also haven't been updated to use the JSON comparator to compare JSON
values.

(cherry picked from commit 0b4507d0d725df83b2bc3bf69fc22c08bfba19f0)