Bug#21808680: JSON + GENERATED COLUMN CORRUPTS TABLE CACHE MEMORY, CRASHES
In some cases pointers to stale data could be left in the Item tree of
the generated column expression, which would lead to problems the next
time the generated column was accessed.

There is already code in place to avoid this situation.
TABLE::cleanup_gc_items() calls Item::cleanup() on the generated
column expression upon completion of a statement, and this is supposed
to revert the Item tree to a pristine state, ready for reuse in a new
statement.

However, since the Item::cleanup() functions are not recursive, only
the top-level item in the generated column expression will be cleaned
up. Children, such as items representing the arguments of a function
call, will not be cleaned up and can point to stale data inserted into
the item tree during execution.

The fix changes TABLE::cleanup_gc_items() so that it calls cleanup()
on all the items in the generated column's item_free_list, and not
only on the top-level item.

The patch also removes an unused parameter from the function.