Bug#21644827 - FTS, ASSERT !SRV_READ_ONLY_MODE || M_IMPL.M_LOG_MODE == MTR_LOG_NO_REDO
Problem:
--------
This bug and Bug#21691208 are because of incorrect table->flags2 value.
For row_format=redundant, we dont read SYS_TABLES.MIX_LEN(flags2)
because we think there could be garbage written before 3.23.52 version.

Fix:
----
We now unconditionally read flags2 from SYS_TABLES and decide if it is
garbage after evaluating if the tablespace is system tablespace, there are no
fts indexes on the table and table is in redundant format

Reviewed-by: Marko Makela <marko.makela@oracle.com>
RB: 10288