BUG#21319192 - CREATING SHARED MEMORY CONNECTIONS WITH A                MULTITHREADED CLIENT HANGS THE CLIENT
The sysbench "run" command hangs when attempting to create multiple tables
for the oltp test using shared memory connections (on Windows).

This problem arises with simultaneous connection attempts using shared memory
due to way in which connection requests are signalled via a pair of
event_connect_request and event_connect_answer Windows events. If multiple
clients set the event_connect_request event more or less simultaneously, the
server may only see a _single_ request and thus only set the
event_connect_answer event once.  This results in a single client getting a
connection, and the rest waiting "forever" (or until connection timeout) for
an event signal that the server will now not send.

The fix is to create global named mutex and serialize the connection event
request from different clients to the server using this mutex. The named mutex
is created in the server as an optimization so that the CreateMutex calls in
client avoid mutex creation costs.