BUG#19930894 - HEAP CORRUPTION AND/OR VALGRIND ERROR IN                CREATE_LENGTH_TO_INTERNAL_LENGTH.
A valgrind error is reported when table is created with
column type as BIT from the data selected from another
table that has a different column type other than BIT type.

For some types pack_flag field is not initialized when
Create_field object is instantiated. This results in the
valgrind error uninitialised value of this variable and
results in the trace shown in the bug report. When
Create_field object is instantiated, initialize pack_flag to
zero. Subsequent table  creation also fails because of
incompatible types and hence log an error.