Bug#21854241: QUERY USING JSON_EXTRACT() RETURNS WRONG RESULT               AFTER ADDING VIRTUAL INDEX
A predicate that compares the result of a JSON function with a string
can be replaced with an indexed generated column compared with the
string. This transformation is not valid, since the generated column
contains double-quoted strings, and the query returns wrong results.

Given the generated column

  gc VARCHAR(100) GENERATED ALWAYS AS (JSON_EXTRACT(...))

and the predicate

  JSON_EXTRACT(...) = 'abc'

it is not correct to substitute the predicate with

  gc = 'abc'

The original predicate compares a JSON value to a string, which means
that the string is automatically converted into a JSON string, and the
predicate will match all rows where JSON_EXTRACT returns the string
"abc".

The rewritten predicate compares a string value to another string
value. Since the generated column contains the result of JSON_EXTRACT
serialized to string, all strings are enclosed in double-quotes. This
means that none of them will match the string literal 'abc', and the
predicate always evaluates to false.

This patch disables the incorrect substitution.