Bug#21514135 SCHEMA MISMATCH ERROR WHEN IMPORTING TABLESPACE AFTER DROPPING AN INDEX
Problem:
========
A schema mismatch error occurs when importing a TABLESPACE from one server
to another. The situation occurs when importing a table has been altered by
DROP INDEX on the source server.

The high level flow is:
1) Drop an index from a table on the source server
2) Move the table to the destination server using transportable TABLESPACES
3) Receive a "SCHEMA MISMATCH" error when importing TABLESPACE

This problem is due to the variable dict_col_t->ord_part of a column, part
of an index to be dropped, not being reset to 0.

Fix:
====
During the dropping of an index, it is checked if the column being part of
the index to be dropped is part of any other index which is not being
dropped. If it is not part of any other index then the dict_col_t->ord_part
variable is reset to 0.

Reviewed-by: Jimmy Yang <Jimmy.Yang@oracle.com>
Reviewed-by: Satya Bodapati <satya.bodapati@oracle.com>
RB: 10212
(cherry picked from commit 6d1717ba576721644bfebaf8c0337973e5736f9f)

Conflicts:
	mysql-test/suite/innodb/r/innodb-alter.result
	mysql-test/suite/innodb/t/innodb-alter.test