Bug #21346081: MEMORY LEAK IN CREATE_TMP_TABLE
Issue:
-----
If an error occurs during the setup phase of the subquery
materialization (used to compute an IN predicate), cleanup
of temporary table does not happen. This creates valgrind
errors.

SOLUTION:
---------
An error can be caused in the setup phase of
subselect_hash_sj_engine, when a fix_fields() triggers the
evaluation of a constant inner subquery because it returned
more than one row. In such a scenario, the object is
deleted. But before doing this, it is important to cleanup
the object since it also contains a temporary table.