Bug#19633824 POINTER TOO LARGE, ERROR 2306, DBDIH (LINE: 17629)
If a data node is known to have been alive at a certain global
checkpoint by other nodes in cluster, but its sysfile reports a
lower GCI, the higher GCI will be used anyway to determine what
global checkpoint that data node will be able to recreate.

This was not ok if starting node have a clean file system
indicated with zero GCI, in which case the higher GCI known by
the other will not be used as by this fix.

Also, it should only be possible for a data node to be at most one
global checkpoint behind according to its sysfile.

This fix only allows that a higher GCI known by the other will be
used if it is at most one GCI ahead.