Bug#20425399: CAN'T USE COLLATE
When a numeric expression is used with a COLLATE clause e.g inside
an ORDER BY, the character set of the expression is set to be
latin1. If the specified collation is associated with a character
set that is incompatible with latin_1, an exception is raised.

Solution: When a numeric expression is implicitly cast to a
character expression by using COLLATE, the character set to be
used for the expression is the same as the one associated with the
named collation.

Reviewed-by: Roy Lyseng <roy.lyseng@oracle.com>
Reviewed-by: Tor Didriksen <tor.didriksen@oracle.com>