Bug#21887035: INCOMPLETE SUPPORT AND/OR DOCUMENTATION OF THE JSON DATATYPE
BLOB, TEXT, GEOMETRY and JSON columns don't support a DEFAULT value.
The error message you get when you try to create such a column with a
DEFAULT value, says that it's a BLOB or TEXT column.

Improve the error message so that it says it's a BLOB, TEXT, GEOMETRY
or JSON column.