Add support to atrt for running concurrent test on same machine.
WL#8736 NDB: Run autotest from Hudson

Atrt used ndb CPC for starting processes on other machines.
Every process started on a host is uniquely identified by a
group and a name in CPC.

For group is the configuration group suffix defined with clusters
argument used, and for name is a sequence number along with
process name used.

If one run several instances of atrt that use same cluster
configuration group, but from different configuration files there
is a conflict.

A site argument are added and if used the CPC group will be a
concatanation of site and cluster group, and so enables concurrent
runs of cluster testing on same machine even with same cluster
groupconfiguration group name.