Bug#21796845 ASSERT ERR == DB_SUCCESS, ROW_QUIESCE_TABLE_START(), FLUSH TABLES FOR EXPORT
Problem:
========
Assertion err == DB_SUCCESS fails during FLUSH TABLES FOR EXPORT for
row_quiesce_table_start().

For non-partitioned table, for flush we set the dict_table_t quiesce state
to QUIESCE_START in ha_innobase::store_lock(), through the function
row_quiesce_set_state(), and we do the flush in ha_innobase::external_lock
and set the state to QUIESCE_COMPLETE, again through row_quiesce_set_state.
Quiesce/Flush is not done if the table is in SYS tablespace.

For partitioned table, the dict_table_t state is set to QUIESCE_START
through row_quiesce_set_state() only for the first partition, in
ha_innobase::store_lock() and we try to quiesce/flush the partitions in
ha_innopart::external_lock() based on the first partition. For all other
partitions we are forcing the state to [table->quiesce=QUIESCE_START],
which is wrong if the partition has SYS tablespace, and hence it hits the
assertion.

Also, for partitioned table, in the case where the first partition is a
SYS tablespace, the partition will not be marked QUIESCE_START and would
skip flushing the remaining partitions.

Fix:
====
Defining ha_innopart::store_lock() and setting quiesce state cleanly for
all partitions through row_quiesce_set_state(), instead of forcing the
state and looping through each of the partitions and flushing it in
ha_innopart::external_lock().

Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>
Reviewed-by: Sunny Bains <Sunny.Bains@oracle.com>
RB: 10374