Bug #11764893 " NDB_RESTORE DOES NOT WARN OF POSSIBLE DUPLICATE KEYS WHEN USING UNIQUE IN"
If ndb_restore is run without --disable-indexes or --rebuild-
indexes, and the schema contains a unique index, the rows may be
restored in a sequence which results in duplicate values for the
unique index. This makes ndb_restore error out with duplicate key
errors.

Added a check to ensure that if --disable-indexes or --rebuild-
indexes is not passed to ndb_restore, it checks all the index
objects and errors out if it finds a unique index.