Patch for bug#21660947
Reseting of 'm_latestGCI' was previously based on detecting
that 'm_active_op_counts' reached zero. Setting this to zero
after an initial restart is essential as the restart resets the GCI
sequence and start over from 0/0 (or close to).

If m_latestGCI is not reset after such an initial restart, we
will reject epochs from after the restart as duplicates of
already completed epochs - see 'bug-subject'.

Depending on the order the client application performs
drop of failed event ops, and re-create of new ones, we might never
reach the state 'm_active_op_count==0'. Thus this can not be
used as a criteria for when to reset m_latestGCI.

This patch reset m_latestGCI *only* as part of handling a CLUSTER_FAILURE.
As part of this failure handling we will already empty
the GCI containers, added comments and asserts about that.
The patch add clearing of m_latestGCI and m_latest_complete_GCI
after handling the failure. There should be no need to distinguise
between initial and non-initial restarts at this point: The
GCI containers will be empty anyway and no more events will arrive
from before the failure.

The init_gci_containers() has been cleaned up to do only that ->
'init the Gci_container', (and the known_gci vector describing its contents)

The previous clearing and reset of 'complete data' has been moved
out to the calle instead, where required.

Note: A CLUSTER_FAILURE must *not* clear the complete date, as that
would have cleared the just created failure event.