Bug#17733850 NAME_CONST() CRASH IN ITEM_NAME_CONST::ITEM_NAME_CONST()
Nested, inside Item_name_const::itemize, we have this piece of code:

  (((Item_func *) value_item)->functype() == Item_func::NEG_FUNC)

Here, the nested NAME_CONST within a NAME_CONST (pointed to by
value_item) as in the repro, is *not* an Item_func, since
Item_name_const is not a subclass of Item_func, but we are led to
believe so because of the special implementation of type() for
Item_name_const: it passes itself off as whatever it nests in its
second argument (Item_name_const::value_item).

Hence trying to invoke Item_func::functype() for Item_name_const
results in an error.

We can not easily detect that value_item is an Item_name_const, since
Item_name_const::type returns the type of its contained value_item -
on the innermost level - even if we had an Item::Type constant
representing NAME_CONST, which we do not.

So, the fix is to add an extra (exceptional use of) dynamic cast to avoid
the NAME_CONST passing itself off as something it's not.

Also, we forbid the use of any other function that could pass as a
literal, e.g. "time '1'", since it never worked anyway (server
exits). The only functions to pass muster are the COLLATE and monadic
NEG functions, anything else will now give an error message.

If there is a use case for other literals implemented as functions, we
can open up for them later in a separate WL, but then we'd need to go
over and check all relevant casts for safety.

I upgraded some C-style casts to down_casts in the code under
scrutiny.  New test cases added.